Fix SQLGetInstalledDrivers single entry error
Fix a bug where SQLGetInstalledDrivers would return an error response
when the "[ODBC Drivers]" section of an odbcinst.ini file contained
only one installed driver entry. The bug was the result of a logical
error where installed driver entries are only copied to the output
buffer if the number of entries was greater than one.